% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testrpackageV2-package.R
\docType{package}
\name{testrpackageV2-package}
\alias{testrpackageV2}
\alias{testrpackageV2-package}
\title{testrpackageV2: Testing an R-package installation}
\description{
This is a test package to learn how to properly install an R-package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggrlab/testrpackagev2}
  \item \url{https://ggrlab.github.io/testrpackageV2/}
  \item Report bugs at \url{https://github.com/ggrlab/testrpackagev2/issues}
}

}
\author{
\strong{Maintainer}: Gunther Glehr \email{gunthergl@gmx.net} (\href{https://orcid.org/0000-0002-1495-9162}{ORCID})

}
\keyword{internal}
